+++ src/services/fetchers/PolygonIOCurrencySnapshotGramsGetter.ts
@@ -10,9 +10,6 @@ import {
 } from '../../types/fetchers.js';
 import {PriceDataRepository} from '../../repositories/PriceDataRepository.js';
 import {PolygonIOCurrencySnapshotGramsDataRepository} from '../../repositories/fetchers/PolygonIOCurrencySnapshotGramsDataRepository.js';
-import {MappingRepository} from '../../repositories/MappingRepository.js';
-import TimeService from '../TimeService.js';
-import {FetchersMappingCacheKeys} from './common/FetchersMappingCacheKeys.js';
 
 export interface PolygonIOCurrencySnapshotGramsInputParams {
   ticker: string;
@@ -27,8 +24,6 @@ export interface PolygonIOCurrencySnapshotGramsInputParams {
 @injectable()
 export class PolygonIOCurrencySnapshotGramsGetter implements FeedFetcherInterface {
   @inject('Logger') protected logger!: Logger;
-  @inject(MappingRepository) private mappingRepository!: MappingRepository;
-  @inject(TimeService) private timeService!: TimeService;
   @inject(PolygonIOCurrencySnapshotGramsDataRepository)
   private pIOCurrencySnapshotGramsDataRepository!: PolygonIOCurrencySnapshotGramsDataRepository;
   @inject(PriceDataRepository) private priceDataRepository!: PriceDataRepository;
@@ -45,12 +40,6 @@ export class PolygonIOCurrencySnapshotGramsGetter implements FeedFetcherInterfac
       return {prices: []};
     }
 
-    try {
-      await this.cacheInput(params);
-    } catch (e) {
-      this.logger.error(`${this.logPrefix} failed cache: ${(e as Error).message}`);
-    }
-
     const {symbols, timestamp} = options;
 
     const prices = await this.pIOCurrencySnapshotGramsDataRepository.getPrices(params, timestamp);
@@ -64,19 +53,5 @@ export class PolygonIOCurrencySnapshotGramsGetter implements FeedFetcherInterfac
     );
 
     return {prices};
-  }
-
-  private async cacheInput(params: PolygonIOCurrencySnapshotGramsInputParams[]): Promise<void> {
-    const timestamp = this.timeService.apply();
-    const key = FetchersMappingCacheKeys.POLYGONIO_CURRENCY_SNAPSHOT_GRAMS_PARAMS;
-
-    const cache = await this.mappingRepository.get(key);
-    const cachedParams = JSON.parse(cache || '{}');
-
-    params.forEach((input) => {
-      cachedParams[input.ticker.toLowerCase()] = timestamp;
-    });
-
-    await this.mappingRepository.set(key, JSON.stringify(cachedParams));
   }
 }