+++ src/services/fetchers/SovrynPriceGetter.ts
@@ -1,5 +1,4 @@
 import {inject, injectable} from 'inversify';
-import {ethers} from 'ethers';
 import {Logger} from 'winston';
 
 import {
@@ -13,9 +12,6 @@ import {
 
 import {PriceDataRepository} from '../../repositories/PriceDataRepository.js';
 import {SovrynDataRepository} from '../../repositories/fetchers/SovrynDataRepository.js';
-import {MappingRepository} from '../../repositories/MappingRepository.js';
-import TimeService from '../TimeService.js';
-import {FetchersMappingCacheKeys} from './common/FetchersMappingCacheKeys.js';
 
 export type SovrynPriceInputParams = {
   base: string;
@@ -38,11 +34,9 @@ weBTC-rUSDT:
 */
 @injectable()
 export class SovrynPriceGetter implements FeedFetcherInterface {
-  @inject(MappingRepository) private mappingRepository!: MappingRepository;
   @inject(SovrynDataRepository) private sovrynDataRepository!: SovrynDataRepository;
   @inject(PriceDataRepository) private priceDataRepository!: PriceDataRepository;
   @inject('Logger') private logger!: Logger;
-  @inject(TimeService) private timeService!: TimeService;
 
   private logPrefix = `[${FetcherName.SovrynPrice}]`;
   static fetcherSource = '';
@@ -51,12 +45,6 @@ export class SovrynPriceGetter implements FeedFetcherInterface {
     this.logger.debug(
       `${this.logPrefix} fetcher started for ${params.map((p) => `[${p.base}/${p.quote}]`).join(', ')}`,
     );
-
-    try {
-      await this.cacheInput(params);
-    } catch (e) {
-      this.logger.error(`${this.logPrefix} failed cache: ${(e as Error).message}`);
-    }
 
     const pricesResponse: FeedPrice[] = await this.sovrynDataRepository.getPrices(params, options.timestamp);
 
@@ -72,28 +60,5 @@ export class SovrynPriceGetter implements FeedFetcherInterface {
     );
 
     return fetcherResult;
-  }
-
-  private async cacheInput(params: SovrynPriceInputParams[]): Promise<void> {
-    const timestamp = this.timeService.apply();
-    const key = FetchersMappingCacheKeys.SOVRYN_PRICE_PARAMS;
-
-    const cache = await this.mappingRepository.get(key);
-    const cachedParams = JSON.parse(cache || '{}');
-
-    params.forEach((input) => {
-      const id = ethers.utils.id(`${input.base};${input.quote}`.toLowerCase());
-
-      cachedParams[id] = {
-        params: <SovrynPriceInputParams>{
-          base: input.base.toLowerCase(),
-          quote: input.quote.toLowerCase(),
-          amountInDecimals: input.amountInDecimals,
-        },
-        timestamp,
-      };
-    });
-
-    await this.mappingRepository.set(key, JSON.stringify(cachedParams));
   }
 }