+++ src/workers/fetchers/MoCMeasurementFetcher.ts
@@ -2,17 +2,17 @@ import axios, {AxiosResponse} from 'axios';
 import {inject, injectable} from 'inversify';
 import {Logger} from 'winston';
 
-import {ServiceInterface} from '../../types/fetchers.js';
+import {FetcherName, ServiceInterface} from '../../types/fetchers.js';
 
 import {
   MoCMeasurementDataRepository,
   MoCMeasurementDataRepositoryInput,
 } from '../../repositories/fetchers/MoCMeasurementDataRepository.js';
 
-import {MappingRepository} from '../../repositories/MappingRepository.js';
-import {FetchersMappingCacheKeys} from '../../services/fetchers/common/FetchersMappingCacheKeys.js';
 import {AxiosResponseChecker} from './_common/AxiosResponseChecker.js';
 import {MoCMeasurementCache} from '../../types/fetchersCachedTypes.js';
+import {DeviationFeedsGetter} from './_common/DeviationFeedsGetter.js';
+import {MoCMeasurementPriceInputParams} from '../../services/fetchers/MoCMeasurementGetter.js';
 
 // field => value
 type ParsedResponse = {measurement_id: string; timestamp: number; field: string; value: number};
@@ -34,8 +34,8 @@ https://api.moneyonchain.com/api/doc
 */
 @injectable()
 export class MoCMeasurementFetcher implements ServiceInterface {
+  @inject(DeviationFeedsGetter) feedsGetter!: DeviationFeedsGetter;
   @inject(AxiosResponseChecker) private axiosResponseChecker!: AxiosResponseChecker;
-  @inject(MappingRepository) private mappingRepository!: MappingRepository;
   @inject(MoCMeasurementDataRepository) private moCMeasurementDataRepository!: MoCMeasurementDataRepository;
   @inject('Logger') private logger!: Logger;
 
@@ -126,17 +126,25 @@ export class MoCMeasurementFetcher implements ServiceInterface {
   }
 
   private async generateInput(): Promise<CachedParams[]> {
-    const paramsKey = FetchersMappingCacheKeys.MOC_MEASUREMENT_PARAMS;
+    const cache = await this.feedsGetter.apply<MoCMeasurementPriceInputParams>(FetcherName.MoCMeasurement);
+    return this.toCacheInput(cache);
+  }
 
-    const cache = await this.mappingRepository.get(paramsKey);
-    const parsed = JSON.parse(cache || '{}') as MoCMeasurementCache;
+  private toCacheInput(inputsParams: MoCMeasurementPriceInputParams[]): CachedParams[] {
+    const idsCache: MoCMeasurementCache = {};
+
+    inputsParams.forEach((input) => {
+      if (!idsCache[input.measurement_id]) idsCache[input.measurement_id] = {};
+
+      idsCache[input.measurement_id][input.field] = true;
+    });
 
     const result: CachedParams[] = [];
 
-    Object.keys(parsed).forEach((measurement_id) => {
+    Object.keys(idsCache).forEach((measurement_id) => {
       result.push({
         measurement_id,
-        fields: Object.keys(parsed[measurement_id]),
+        fields: Object.keys(idsCache[measurement_id]),
       });
     });